Fix brush.move during a brush gesture.
The brush captures the selection at the start of a brush gesture. If the
selection is set programatically by brush.move during a brush gesture, the
captured reference can be different than the brush’s new selection! Now we
always use the element’s current brush selection.

In addition, the brush no longer modifies the selection in-place. If a caller
keeps a reference to the brush selection, it will no longer be affected.